+++ Data/Models/Order.cs
@@ -4,11 +4,11 @@
     {
         public int OrderId { get; set; }
         public Guid UserId { get; set; }
-        public double SubTotalAmount { get; set; }
+        public int SubTotalAmount { get; set; }
         public TrolleyMethod ServiceMethod { get; set; }
-        public double BagFee { get; set; }
-        public double ServiceFee { get; set; }
-        public double TotalAmount { get; set; }
+        public int BagFee { get; set; }
+        public int ServiceFee { get; set; }
+        public int TotalAmount { get; set; }
         public string ChosenAddress { get; set; }
         public DateTime ChosenDate { get; set; }
         public OrderStatus OrderStatus { get; set; }
@@ -20,7 +20,7 @@
     {
         public int OrderId { get; set; }
         public TrolleyMethod ServiceMethod { get; set; }
-        public double TotalAmount { get; set; }
+        public int TotalAmount { get; set; }
         public string ChosenAddress { get; set; }
         public DateTime ChosenDate { get; set; }
         public OrderStatus OrderStatus { get; set; }
@@ -32,22 +32,22 @@
         public int OrderItemId { get; set; }
         public int ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public int Price { get; set; }
         public PriceUnitType PriceUnitType { get; set; } = PriceUnitType.ea;
         public string? Image { get; set; }
-        public float Quantity { get; set; }
-        public float TotalPrice { get; set; }
+        public int Quantity { get; set; }
+        public int TotalPrice { get; set; }
     }
 
     public class OrderContent
     {
         public int OrderId { get; set; }
         public List<OrderItem> OrderItems { get; set; } = new List<OrderItem>();
-        public double SubTotalAmount { get; set; }
+        public int SubTotalAmount { get; set; }
         public TrolleyMethod ServiceMethod { get; set; }
-        public double BagFee { get; set; }
-        public double ServiceFee { get; set; }
-        public double TotalAmount { get; set; }
+        public int BagFee { get; set; }
+        public int ServiceFee { get; set; }
+        public int TotalAmount { get; set; }
         public string ChosenAddress { get; set; }
         public DateTime ChosenDate { get; set; }
         public OrderStatus OrderStatus { get; set; }
